Forge fmt

+++ src/DecentralizedCollegeTracker.sol
@@ -20,113 +20,68 @@ contract DecentralizedUniversity {
         address cAdmin;
         string cRegNo;
         bool isAllowedToAddStudents;
-        uint totalNoOfStudents;
+        uint256 totalNoOfStudents;
     }
 
     struct Student {
         string sName;
-        uint phoneNo;
+        uint256 phoneNo;
         string courseEnrolled;
     }
 
     mapping(address => College) colleges; //  Mapping a college's address to college
     mapping(string => Student) students; //  Mapping a student's name to student
 
-    function addNewCollege(
-        string memory collegeName,
-        address add,
-        address cAdmin,
-        string memory regNo
-    ) public onlyAdmin {
-        require(
-            !areBothStringSame(colleges[add].cName, collegeName),
-            "College already exists with same name"
-        );
+    function addNewCollege(string memory collegeName, address add, address cAdmin, string memory regNo)
+        public
+        onlyAdmin
+    {
+        require(!areBothStringSame(colleges[add].cName, collegeName), "College already exists with same name");
         colleges[add] = College(collegeName, add, cAdmin, regNo, true, 0);
         totalNoOfColleges++;
     }
 
-    function viewCollegeDetails(
-        address add
-    ) public view returns (string memory, string memory, uint) {
-        return (
-            colleges[add].cName,
-            colleges[add].cRegNo,
-            colleges[add].totalNoOfStudents
-        );
+    function viewCollegeDetails(address add) public view returns (string memory, string memory, uint256) {
+        return (colleges[add].cName, colleges[add].cRegNo, colleges[add].totalNoOfStudents);
     }
 
     function blockCollegeToAddNewStudents(address add) public onlyAdmin {
-        require(
-            colleges[add].isAllowedToAddStudents,
-            "College is already blocked to add new students"
-        );
+        require(colleges[add].isAllowedToAddStudents, "College is already blocked to add new students");
         colleges[add].isAllowedToAddStudents = false;
     }
 
     function unblockCollegeToAddNewStudents(address add) public onlyAdmin {
-        require(
-            !colleges[add].isAllowedToAddStudents,
-            "College is already unblocked to add new students"
-        );
+        require(!colleges[add].isAllowedToAddStudents, "College is already unblocked to add new students");
         colleges[add].isAllowedToAddStudents = true;
     }
 
-    function addNewStudentToCollege(
-        address add,
-        string memory sName,
-        uint phoneNo,
-        string memory courseName
-    ) public {
-        require(
-            colleges[add].isAllowedToAddStudents,
-            "This College is blocked to add new students"
-        );
-        require(
-            colleges[add].cAdmin == msg.sender,
-            "Only College admin can add the new student"
-        );
+    function addNewStudentToCollege(address add, string memory sName, uint256 phoneNo, string memory courseName)
+        public
+    {
+        require(colleges[add].isAllowedToAddStudents, "This College is blocked to add new students");
+        require(colleges[add].cAdmin == msg.sender, "Only College admin can add the new student");
         students[sName] = Student(sName, phoneNo, courseName);
         colleges[add].totalNoOfStudents += 1;
         totalNoOfStudents++;
     }
 
-    function getNumberOfStudentForCollege(
-        address add
-    ) public view returns (uint) {
+    function getNumberOfStudentForCollege(address add) public view returns (uint256) {
         return (colleges[add].totalNoOfStudents);
     }
 
-    function viewStudentDetails(
-        string memory sName
-    ) public view returns (string memory, uint, string memory) {
-        return (
-            students[sName].sName,
-            students[sName].phoneNo,
-            students[sName].courseEnrolled
-        );
+    function viewStudentDetails(string memory sName) public view returns (string memory, uint256, string memory) {
+        return (students[sName].sName, students[sName].phoneNo, students[sName].courseEnrolled);
     }
 
-    function changeStudentCourse(
-        address add,
-        string memory sName,
-        string memory newCourse
-    ) public {
+    function changeStudentCourse(address add, string memory sName, string memory newCourse) public {
         require(
-            !areBothStringSame(students[sName].courseEnrolled, newCourse),
-            "Student already enrolled to same course"
+            !areBothStringSame(students[sName].courseEnrolled, newCourse), "Student already enrolled to same course"
         );
-        require(
-            colleges[add].cAdmin == msg.sender,
-            "Only College admin can change the student course"
-        );
+        require(colleges[add].cAdmin == msg.sender, "Only College admin can change the student course");
         students[sName].courseEnrolled = newCourse;
     }
 
-    function areBothStringSame(
-        string memory a,
-        string memory b
-    ) private pure returns (bool) {
+    function areBothStringSame(string memory a, string memory b) private pure returns (bool) {
         if (bytes(a).length != bytes(b).length) {
             return false;
         } else {
